pkg_pkg: allow suppression of spurious package dependencies
Allow suppression of spurious automatic package dependencies by
introducing a new variable PKG_REQUIRES_SUPPRESS, containing a
single-line string of space-separated package list.  Any of the
packages listed are automatically determined to be a requirement are
suppressed so that they do not show up in the generated hint file(s).